requirements.txt
scapy==2.7.0
pytest==8.0.0
allure-pytest==2.13.2
